speedup binary build
- 70s to 20s
